src/main/java/racingcar/Car.java
package racingcar;

public class Car {
    public static final int MAX_NAME_LENGTH = 5;
    public static final int CAN_MOVE_BOUND = 4;

    private final String name;
    private int position = 1;

    public Car(String name) {
        validateCarName(name);
        this.name = name;
    }

    private static void validateCarName(String name) {
        if (name == null || name.isEmpty()) {
            throw new IllegalArgumentException();
        }
        if (name.length() > MAX_NAME_LENGTH) {
            throw new IllegalArgumentException();
        }
    }

    public String getName() {
        return name;
    }

    public int getPosition() {
        return position;
    }

    public void move(int randomNumber) {
        if (randomNumber >= CAN_MOVE_BOUND) {
            position++;
        }
    }

    @Override
    public String toString() {
        return name + " : " + "-".repeat(position);
    }
}
